feat: started adding SpringSecurity (notebook)

+++ src/main/java/com/SpringSecurityDemo/SpringSecurity/domain/product/Product.java
@@ -10,7 +10,7 @@ public class Product {
     @GeneratedValue(strategy =  GenerationType.UUID)
     private String id;
     private String name;
-    private Integer price;
+    private Double price;
     private Integer quantity;
     private Double cost;
 
@@ -23,13 +23,13 @@ public class Product {
 
     public String getId() {return id;}
     public String getName() {return name;}
-    public Integer getPrice() {return price;}
+    public Double getPrice() {return price;}
     public Integer getQuantity() {return quantity;}
     public Double getCost() {return cost;}
 
     public void setId(String id) {this.id = id;}
     public void setName(String name) {this.name = name;}
-    public void setPrice(Integer price) {this.price = price;}
+    public void setPrice(Double price) {this.price = price;}
     public void setQuantity(Integer quantity) {this.quantity = quantity;}
     public void setCost(Double cost) {this.cost = cost;}
